Fix interger overflow on 32bit systems.

+++ FileZilla3/trunk/src/engine/timeex.cpp
@@ -89,7 +89,7 @@ CDateTime CDateTime::Now()
 	CDateTime ret;
 	timeval tv = { 0, 0 };
 	if (gettimeofday(&tv, 0) == 0) {
-		ret.t_ = tv.tv_sec * 1000 + tv.tv_usec / 1000;
+		ret.t_ = static_cast<int64_t>(tv.tv_sec) * 1000 + tv.tv_usec / 1000;
 		ret.a_ = milliseconds;
 	}
 	return ret;
